requirements.txt
pymongo[srv]
motor
fastapi
uvicorn
retrying
pillow
python-dotenv
python-multipart
